main: move cloth input area out of root
This commit alters the structure of the application's input tree so that
the pointer input area used for input to the cloth is not the root of the
input tree. Previously, as the root of the tree, it would receive copies of
all pointer input events routed to its children if they matched a type that
it subscribed to. Moving it into a separate branch of the tree (by wrapping
it in a new clipping area that is popped before laying out other UI) ensures
that events from other UI do not propagate up the tree to it.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>